+++ src/components/RegularButton.tsx
@@ -14,7 +14,7 @@ interface RegularButtonProps {
 function RegularButton({ children, handleClick, handleNextLevel }: RegularButtonProps) {
   return (
     <button
-      className="bg-slate-950 p-4 border-2  border-sky-100 rounded-xl cursor-pointer text-teal-50 text-2xl w-max m-0 mx-auto py-3 px-8 hover:bg-pink-950 focus:bg-red-950"
+      className="bg-slate-950 p-4 border-2  border-sky-100 rounded-xl cursor-pointer text-teal-50 text-2xl w-max m-0 mx-auto py-4 px-17 sm:py-3 sm:px-8 hover:bg-pink-950 focus:bg-red-950"
       onClick={(e) => {
         if(handleClick) {
           handleClick(e);
